Removes python 3.2 from travis.yml
Python 3.2 has been removed as it does not support specifying unicode literals as u'literal'
and throws a syntax error. This prevents the twitter api from working. Since finding a fix would
require meddling with the api code (which as I understand is out of the scope of this project), the
solution found was to remove python 3.2 support.

python versions from  3.3 works fine as the support for u'literal' has been restored
https://www.python.org/dev/peps/pep-0414/